requirements.txt
streamlit
langchain
faiss-cpu
pymupdf
python-dotenv
langchain-community
langchain-google-genai